do not use `new Date()`
exceptions are isolated to utility functions
