+++ web-client/src/presenter/computeds/formattedCaseDetail.test.js
@@ -943,7 +943,7 @@ describe('formattedCaseDetail', () => {
       };
       const caseDeadlines = [
         {
-          deadlineDate: new Date(),
+          deadlineDate: applicationContext.getUtilities().createISODateString(),
         },
       ];
